Extract definitions into ApiTask.ts and models.ts
