+++ websites/rushstack.io/src/pages/community/events.tsx
@@ -4,12 +4,9 @@ import { CommunitySidebarLayout } from "../../rscommunity/view/CommunitySidebarL
 import { CommunitySignInPage } from "../../rscommunity/view/CommunitySignInPage";
 import { AppSession } from "../../rscommunity/api/AppSession";
 import { EventCard } from "../../rscommunity/view/EventCard";
-import {
-  EventModel,
-  ApiTask,
-  ApiTaskStatus,
-} from "../../rscommunity/api/ApiDataService";
 import { ObjectEvent } from "../../rscommunity/library/ObjectEvent";
+import { EventModel } from "../../rscommunity/api/models";
+import { ApiTask, ApiTaskStatus } from "../../rscommunity/api/ApiTask";
 
 class EventsPage extends React.Component {
   private _appSession: AppSession;
